fix handling of curly braces for choice's callback
The assertion `Choice` provided by Symfont accept an option `callback`
which can either be of type `string|array|Closure`.
In this case I think we are only interested by the `array` case since
it's the only one in which a class name can be provided.

The phpdoc parser will provide us with a node of type `CurlyListNode`
but it's currently treated like an array.
In addition the `AbstractValuesAwareNode::changeValue` method will
always check if the new value is quoted or not even when the value is
not actually a `string`.