[Downgrade PHP 7.0] Add Exception fallback for instanceof Throwable
PHP 7.0 introduced Throwable and changed the exception hierarchy,
adding Throwable as its new root. This means various error handlers
will now check for Throwable:

    $e instanceof Throwable

To preserve compatibility with PHP 5.6, we need to check for both
Throwable and Exception. And since the value can possibly produce side-effects,
we need to assign it to a temporary variable:

    (($throwable = $loggedMessage->getExtraObject()) instanceof Throwable || $throwable instanceof \Exception)