DependencyClassMethodDecorator: Prevent duplication of arguments
When a parent class has a constructor-injected dependency and `ClassDependencyManipulator`
tries to add the same dependency to the constructor of the child class,
the child constructor would have two parameters with the same name.

Let’s skip the arguments whose all characteristics (name, type, byRefness, variadicity)
match one of the existing arguments of the child constructor.

In the future, we might also want to handle duplication of parameters
with non-matching characteristics by renaming the parameters.

Fixes: https://github.com/rectorphp/rector/issues/7286