Add failing test fixture for RemoveDeadStmtRector
# Failing Test for RemoveDeadStmtRector

Based on https://getrector.org/demo/a0eea85b-e17c-4bd5-851f-ba654335e096

Not sure if no change at all in this case or it should be removed. Ideally it would give an error, but that's not an option with rector. So I guess completely removing it maybe without leaving broken code?