Add parent constructor call only once
If a promoted property is added to a class constructor and if this constructor method is already present in the class, there is no need to add a parent constructor call to this method just because a new promoted property is added to the child's constructor. Adding the parent constructor call is only necessary when the child class currently has no constructor when the promoted property is added so that a new constructor method has to be added to the child class.

Fixes rectorphp/rector-symfony#275