fix: ArrayToFirstClassCallableRector should convert private in own scope
If an array callable is explicitly referencing `$this`, then it should
always be safe to convert it to a first-class callable even if the
method is private / protected.

This was the original behaviour of the 8.1 fixers, but it changed in
v1.1.1 following a fix for cases where a private / protected method was
being referenced from outside the owning scope.