[Php81] ArrayToFirstClassCallableRector should not skip non-public methods from owning scope (#7760)
* test: Prove ArrayToFirstClassCallableRector skips internal callbacks

If a class contains array callbacks with a non-public method, they are
not converted into first-class callables. This is due to a fix for a bug
that was specific to skipping non-public methods referenced from
*outside* the owning class - there was no previous coverage for what
should happen when a callable is referenced within a valid scope.

* fix: ArrayToFirstClassCallableRector should convert private in own scope

If an array callable is explicitly referencing `$this`, then it should
always be safe to convert it to a first-class callable even if the
method is private / protected.

This was the original behaviour of the 8.1 fixers, but it changed in
v1.1.1 following a fix for cases where a private / protected method was
being referenced from outside the owning scope.

* Apply fixture class naming changes from code review

Co-authored-by: Abdul Malik Ikhsan <samsonasik@gmail.com>

---------

Co-authored-by: Abdul Malik Ikhsan <samsonasik@gmail.com>